+++ assets/js/backend.js
@@ -1,1 +1,4 @@
 import '../css/backend.scss';
+
+// start the Stimulus applications
+import '../stimulus_bootstrap';