+++ src/Migrations/Version20210913091421.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DoctrineMigrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20210913091421 extends AbstractMigration
+{
+    public function getDescription() : string
+    {
+        return '';
+    }
+
+    public function up(Schema $schema) : void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('CREATE TABLE events_students (event_id INT NOT NULL, student_id INT NOT NULL, has_attended_the_class TINYINT(1) DEFAULT \'1\' NOT NULL, INDEX IDX_595D15B171F7E88B (event_id), INDEX IDX_595D15B1CB944F1A (student_id), UNIQUE INDEX UNIQ_DB5678D3EA12567 (event_id, student_id), PRIMARY KEY(event_id, student_id)) DEFAULT CHARACTER SET utf8mb4 COLLATE `utf8mb4_unicode_ci` ENGINE = InnoDB');
+        $this->addSql('ALTER TABLE events_students ADD CONSTRAINT FK_595D15B171F7E88B FOREIGN KEY (event_id) REFERENCES event (id)');
+        $this->addSql('ALTER TABLE events_students ADD CONSTRAINT FK_595D15B1CB944F1A FOREIGN KEY (student_id) REFERENCES student (id)');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('DROP TABLE events_students');
+    }
+}